requirements.txt
alembic==1.0.0
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
cffi==1.11.5
click==6.7
cryptography==2.3
cycler==0.10.0
Flask==1.0.2
Flask-BasicAuth==0.2.0
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
joblib==0.12.2
kiwisolver==1.0.1
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
meld3==1.0.2
nltk==3.3
numpy==1.15.0
pandas==0.23.4
pycparser==2.18
PyMySQL==0.9.2
pyparsing==2.2.0
PySastrawi==1.1.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
requests==2.19.1
scipy==1.1.0
scikit-learn==0.19.2
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.10
tzlocal==1.5.1
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
